+++ src/MediaControl/CuePoint.js
@@ -117,23 +117,49 @@ CuePoints.contextTypes = {
 }
 
 const enhance = compose(
-  getContext({ currentSong: PropTypes.object }, (props) => ({ currentSong: props.currentSong })),
+  getContext({
+    currentSong: PropTypes.object,
+    store: PropTypes.object
+  }, (props) => ({
+    currentSong: props.currentSong,
+    store: props.store
+  })),
   lifecycle({
     componentDidMount() {
       if(typeof window === 'undefined') return
 
       document.addEventListener('keydown', (ev) => {
         const KEY_1 = 49
-        if(ev.which < KEY_1 || ev.which > KEY_1+8) return
+        const KEY_L = 76
+        const KEY_N = 78
 
-        const codeKey = ev.which - KEY_1
-        const { cuePoints } = this.props.currentSong
+        if(ev.which >= KEY_1 && ev.which <= KEY_1+8) {
 
-        if(codeKey >= cuePoints.length) return
+          const codeKey = ev.which - KEY_1
+          const { cuePoints } = this.props.currentSong
 
-        const cuePoint = cuePoints[codeKey]
-        const seekTo = cuePoint.start / this.props.wavesurfer.getDuration()
-        this.props.wavesurfer.seekTo(seekTo)
+          if(codeKey >= cuePoints.length) return
+          const cuePoint = cuePoints[codeKey]
+
+          if(ev.shiftKey) {
+            deleteCuePoint(this.props.store, cuePoint, this.props.wavesurfer)
+            return
+          }
+
+          const seekTo = cuePoint.start / this.props.wavesurfer.getDuration()
+          this.props.wavesurfer.seekTo(seekTo)
+          return
+        }
+
+        if(ev.which === KEY_L) {
+          toggleActive(this.props.store, this.props.wavesurfer)
+          return
+        }
+
+        if(ev.which === KEY_N) {
+          addLoop(this.props.store, this.props.wavesurfer, this.props.currentSong)
+          return
+        }
       })
     }
   })